+++ src/helpers/drawing.ts
@@ -5,13 +5,14 @@ import {
   ImgState,
   changeSvgState,
 } from '../images/actions/images';
-import { formatPrice } from './parser';
+import { formatPrice, addThousandSeperator } from './parser';
 
 export async function drawQuoteImage(
   ticker: string,
   icon: string,
   price: number,
   percentage: number,
+  totalValue: number,
   colors: { increasing: ImgState; decreasing: ImgState }
 ): Promise<string> {
   const canvasWidth = 144;
@@ -24,11 +25,16 @@ export async function drawQuoteImage(
   ctx.strokeStyle = '#fff'; // Set the divider line color
   ctx.lineWidth = 3; // Set the divider line width
 
-  // Draw the divider line
-  ctx.beginPath();
-  ctx.moveTo(10, 61);
-  ctx.lineTo(67, 61);
-  ctx.stroke();
+  let offsetY = -24;
+
+  if(totalValue === 0) {
+    // Draw the divider line
+    ctx.beginPath();
+    ctx.moveTo(10, 61);
+    ctx.lineTo(67, 61);
+    ctx.stroke();
+    offsetY = 0;
+  }
 
   // Determine color and SVG content based on price change
   let svgContent = percentage >= 0 ? arrowUp : arrowDown;
@@ -71,18 +77,25 @@ export async function drawQuoteImage(
   ctx.font = `500 ${pSize}pt "Verdana"`;
   const priceText = `${formattedPrice}`;
   const textWidth = ctx.measureText(priceText).width;
-  ctx.fillText(priceText, 10, 98);
+  ctx.fillText(priceText, 10, 98 + offsetY);
 
   // Draw the arrow icon next to the price
   const arrowSize = formattedPrice.length > priceLength ? 15 : 18; // Set arrow size
   const aPosition = formattedPrice.length > priceLength ? 82 : 80; // Adjust arrow position
   const arrowX = 10 + textWidth + 5; // X position to draw the arrow (right of the price text)
-  ctx.drawImage(arrowImage, arrowX, aPosition, arrowSize, arrowSize); // Adjust positioning as needed
+  ctx.drawImage(arrowImage, arrowX, aPosition + offsetY, arrowSize, arrowSize); // Adjust positioning as needed
+
+  offsetY = totalValue ? offsetY - 3 : offsetY;
 
   // Set the fill color for percentage change
   ctx.fillStyle = state;
   ctx.font = 'normal 16pt "Verdana"';
-  ctx.fillText(`(${percentage.toFixed(2)}%)`, 10, 128);
+  ctx.fillText(`(${percentage.toFixed(2)}%)`, 10, 128 + offsetY);
+
+  if(totalValue) {
+    ctx.fillStyle = 'white';
+    ctx.fillText(addThousandSeperator(Math.round(totalValue)), 10, 128);
+  }
 
   return canvas.toDataURL();
 }